Simplify NodeDiscoery
