+++ crates/utils/sov-proxy-utils/src/node_discovery.rs
@@ -1,0 +1,438 @@
+//! Utilities for discovering cluster nodes from PostgreSQL and persisting snapshots.
+use crate::file_writer::write_to_file_atomically;
+use anyhow::{Context, Result};
+use async_trait::async_trait;
+use sqlx::postgres::{PgListener, PgPool};
+use sqlx::FromRow;
+use std::collections::btree_map::Entry;
+use std::collections::{BTreeMap, BTreeSet};
+use std::net::SocketAddr;
+use std::path::PathBuf;
+use std::time::Duration;
+pub use time::OffsetDateTime;
+use tokio::sync::watch;
+use tokio::task::JoinHandle;
+
+/// Information about a registered node.
+#[derive(Debug, Clone, FromRow, PartialEq, Eq)]
+pub struct NodeInfo {
+    /// The unique identifier of the node.
+    pub node_id: String,
+    /// The address (ip:port) where the node can be reached.
+    pub address: SocketAddr,
+    /// The last time the node updated its heartbeat.
+    pub last_updated: OffsetDateTime,
+}
+
+impl NodeInfo {
+    fn to_file_line(&self, role: &str) -> String {
+        let ts = self
+            .last_updated
+            .format(&time::format_description::well_known::Rfc3339)
+            .unwrap_or_else(|e| panic!("Failed to format timestamp: {e}"));
+
+        format!("{role}={},{ts},{}", self.address, self.node_id)
+    }
+}
+
+/// Result of querying cluster node information.
+#[derive(Debug, Clone, Default)]
+pub struct ClusterInfo {
+    /// The leader node, if one exists.
+    pub leader: Option<NodeInfo>,
+    /// All follower nodes.
+    pub followers: BTreeMap<String, NodeInfo>,
+}
+
+impl ClusterInfo {
+    /// Returns true if a leader with the given node_id exists.
+    pub fn has_leader(&self, node_id: &str) -> bool {
+        self.leader.as_ref().is_some_and(|l| l.node_id == node_id)
+    }
+
+    /// Returns true if a follower with the given node_id exists.
+    pub fn has_follower(&self, node_id: &str) -> bool {
+        self.followers.contains_key(node_id)
+    }
+
+    /// Formats the cluster info as a string suitable for writing to a file.
+    /// Each node is written on a separate line in the format `role=address,timestamp,node_id`.
+    pub fn to_file_content(&self) -> String {
+        let mut lines = Vec::new();
+
+        if let Some(leader) = &self.leader {
+            lines.push(leader.to_file_line("leader"));
+        }
+
+        for follower in self.followers.values() {
+            lines.push(follower.to_file_line("follower"));
+        }
+
+        lines.join("\n")
+    }
+
+    fn leader_id(&self) -> Option<String> {
+        self.leader.as_ref().map(|l| l.node_id.clone())
+    }
+}
+
+impl NodeInfo {
+    /// Parses a node info from a string in the format `address,timestamp,node_id`.
+    pub fn parse(s: &str) -> Result<Self> {
+        let mut parts = s.split(',');
+        let addr = parts.next().context("Missing address")?;
+        let ts = parts.next().context("Missing timestamp")?;
+        let node_id = parts.next().context("Missing node_id")?;
+
+        Ok(NodeInfo {
+            node_id: node_id.to_string(),
+            address: addr.parse().context("Invalid address")?,
+            last_updated: OffsetDateTime::parse(ts, &time::format_description::well_known::Rfc3339)
+                .context("Invalid timestamp")?,
+        })
+    }
+}
+
+/// Trait for receiving notifications when cluster state changes.
+///
+/// Implement this trait to define custom behavior when the cluster membership
+/// or leader changes. The notification is triggered after the cluster info file
+/// has been updated.
+#[async_trait]
+pub trait ClusterUpdateNotifier: Send + Sync + 'static {
+    /// Called when cluster membership or leadership changes.
+    async fn on_cluster_update(&self, cluster_info: &ClusterInfo);
+}
+
+/// A notifier that intentionally does nothing.
+pub struct NoOp;
+
+#[async_trait]
+impl ClusterUpdateNotifier for NoOp {
+    async fn on_cluster_update(&self, _cluster_info: &ClusterInfo) {}
+}
+
+/// Handle returned when subscribing to cluster updates.
+pub struct NodeDiscoveryTask {
+    pub receiver: watch::Receiver<ClusterInfo>,
+    pub handle: JoinHandle<anyhow::Result<()>>,
+}
+
+/// Client for querying cluster information from the database.
+pub struct NodeDiscovery {
+    max_age: Duration,
+    pool: PgPool,
+    listener: PgListener,
+    prev_followers: BTreeSet<String>,
+    prev_leader_id: Option<String>,
+    path: PathBuf,
+    notifier: Box<dyn ClusterUpdateNotifier>,
+}
+
+impl NodeDiscovery {
+    /// Creates a new NodeDiscovery with a connection pool and custom max_age.
+    ///
+    /// The `max_age` parameter controls how long a node can go without updating
+    /// its heartbeat before being filtered out of the cluster info.
+    ///
+    /// Returns a [`NodeDiscovery`] ready to subscribe for cluster updates.
+    pub async fn connect(
+        connection_string: &str,
+        max_age: Duration,
+        path: PathBuf,
+        notifier: Box<dyn ClusterUpdateNotifier>,
+    ) -> Result<Self> {
+        tracing::info!("Connecting to database.");
+        let pool = sqlx::postgres::PgPoolOptions::new()
+            .max_connections(5)
+            .connect(connection_string)
+            .await?;
+        tracing::info!("DB connection established.");
+
+        // On startup, write an empty file. If the cluster is not empty, the file will be populated on the first call to `handle_cluster_update`.
+        write_to_file_atomically(&path, "").await?;
+
+        let mut listener = PgListener::connect(connection_string).await?;
+        listener
+            .listen_all(["nodes_changes", "leader_changes"])
+            .await?;
+        tracing::info!("Subscribed to nodes_changes and leader_changes channels");
+
+        Ok(Self {
+            max_age,
+            pool,
+            listener,
+            prev_followers: BTreeSet::new(),
+            prev_leader_id: None,
+            path,
+            notifier,
+        })
+    }
+
+    // Fetches cluster info atomically using a single transaction.
+    // Only nodes whose `last_updated` timestamp is within `max_age` are returned.
+    //
+    // # Returns
+    // A `ClusterInfo` struct containing the leader and all followers.
+    async fn get_cluster_info(&self) -> anyhow::Result<ClusterInfo> {
+        let (leader_id, all_nodes) = self.get_cluster_info_from_db().await?;
+        Self::cluster(leader_id, all_nodes)
+    }
+
+    fn cluster(
+        leader_id: Option<String>,
+        all_nodes: Vec<(String, String, OffsetDateTime)>,
+    ) -> anyhow::Result<ClusterInfo> {
+        let mut leader = None;
+        let mut followers = BTreeMap::new();
+
+        for (node_id, address, last_updated) in all_nodes {
+            let address: SocketAddr = address.parse()?;
+            let node = NodeInfo {
+                node_id,
+                address,
+                last_updated,
+            };
+
+            if Some(&node.node_id) == leader_id.as_ref() {
+                leader = Some(node);
+            } else {
+                match followers.entry(node.node_id.clone()) {
+                    Entry::Vacant(vacant_entry) => {
+                        vacant_entry.insert(node);
+                    }
+                    Entry::Occupied(occupied) => {
+                        anyhow::bail!("Duplicate node id found in Nodes table: {}", occupied.key());
+                    }
+                }
+            }
+        }
+
+        if let Some(leader_id) = &leader_id {
+            if leader.is_none() {
+                anyhow::bail!("Leader is missing from the Nodes table. leader_id: {leader_id}");
+            }
+        }
+
+        Ok(ClusterInfo { leader, followers })
+    }
+
+    /// Subscribes to PostgreSQL notifications for cluster changes and writes
+    /// cluster info to a file whenever the cluster state changes.
+    ///
+    /// Listens on `nodes_changes` and `leader_changes` channels.
+    pub async fn spawn(mut self) -> NodeDiscoveryTask {
+        let (sender, receiver) = watch::channel(ClusterInfo::default());
+        let handle = tokio::spawn(async move {
+            loop {
+                if let Err(error) = self.handle_cluster_update(&sender).await {
+                    tracing::warn!(?error, "Cluster update failed");
+                    tokio::time::sleep(Duration::from_millis(1000)).await;
+                }
+            }
+        });
+
+        NodeDiscoveryTask { receiver, handle }
+    }
+
+    async fn handle_cluster_update(
+        &mut self,
+        sender: &watch::Sender<ClusterInfo>,
+    ) -> anyhow::Result<()> {
+        let info = self.get_cluster_info().await?;
+        let leader_id = info.leader_id();
+
+        let followers: BTreeSet<_> = info.followers.keys().cloned().collect();
+
+        let membership_changed = self.prev_followers != followers;
+        let leader_changed = self.prev_leader_id != leader_id;
+
+        tracing::trace!(info = ?info, "Last cluster info");
+
+        if membership_changed || leader_changed {
+            // Log membership changes
+            for node_id in followers.difference(&self.prev_followers) {
+                tracing::info!(node_id, "Node joined the cluster");
+            }
+            for node_id in self.prev_followers.difference(&followers) {
+                tracing::info!(node_id, "Node left the cluster");
+            }
+
+            // Log leader change
+            if leader_changed {
+                tracing::info!(
+                    old_leader = ?self.prev_leader_id,
+                    new_leader = ?leader_id,
+                    "Leader changed"
+                );
+            }
+
+            let content = info.to_file_content();
+            write_to_file_atomically(&self.path, &content).await?;
+            tracing::info!(?self.path, content, "Cluster info file updated");
+
+            self.prev_followers = followers;
+            self.prev_leader_id = leader_id;
+
+            // Notify watchers that the cluster was updated.
+            self.notifier.on_cluster_update(&info).await;
+            let _ = sender.send(info);
+        }
+
+        // Wait for at least one notification.
+        self.listener.recv().await?;
+
+        // Drain any additional pending notifications.
+        while self.listener.next_buffered().is_some() {}
+
+        Ok(())
+    }
+
+    async fn get_cluster_info_from_db(
+        &self,
+    ) -> Result<(Option<String>, Vec<(String, String, OffsetDateTime)>)> {
+        let max_age_secs = self.max_age.as_secs() as i64;
+
+        // Fetch nodes updated within max_age, always including the leader regardless of age.
+        // The leader_id is included in each row via LEFT JOIN, allowing us to get it from the results.
+        let rows: Vec<(String, String, OffsetDateTime, Option<String>)> = sqlx::query_as(
+            "SELECT n.node_id, n.address, n.last_updated, l.node_id as leader_id \
+             FROM nodes n \
+             LEFT JOIN sequencer_leader l ON l.singleton = 1 \
+             WHERE n.last_updated > NOW() - $1 * INTERVAL '1 second' \
+                OR n.node_id = l.node_id \
+             ORDER BY n.node_id",
+        )
+        .bind(max_age_secs)
+        .fetch_all(&self.pool)
+        .await?;
+
+        // Extract leader_id from first row (same in all rows due to LEFT JOIN)
+        let leader_id = rows.first().and_then(|(_, _, _, lid)| lid.clone());
+
+        // Convert to the expected format without leader_id column
+        let all_nodes = rows
+            .into_iter()
+            .map(|(node_id, address, last_updated, _)| (node_id, address, last_updated))
+            .collect();
+
+        Ok((leader_id, all_nodes))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn test_timestamp() -> OffsetDateTime {
+        OffsetDateTime::UNIX_EPOCH
+    }
+
+    const TEST_TIME_STAMP: &str = "1970-01-01T00:00:00Z";
+
+    #[test]
+    fn cluster_with_leader_and_followers() {
+        let ts = test_timestamp();
+        let info = NodeDiscovery::cluster(
+            Some("node1".to_string()),
+            vec![
+                ("node1".to_string(), "127.0.0.1:8000".to_string(), ts),
+                ("node2".to_string(), "127.0.0.1:8001".to_string(), ts),
+                ("node3".to_string(), "127.0.0.1:8002".to_string(), ts),
+            ],
+        )
+        .unwrap();
+
+        assert_eq!(
+            info.to_file_content(),
+            format!(
+                "leader=127.0.0.1:8000,{TEST_TIME_STAMP},node1\n\
+                 follower=127.0.0.1:8001,{TEST_TIME_STAMP},node2\n\
+                 follower=127.0.0.1:8002,{TEST_TIME_STAMP},node3"
+            ),
+        );
+    }
+
+    #[test]
+    fn cluster_with_no_leader() {
+        let ts = test_timestamp();
+        let info = NodeDiscovery::cluster(
+            None,
+            vec![
+                ("node1".to_string(), "127.0.0.1:8000".to_string(), ts),
+                ("node2".to_string(), "127.0.0.1:8001".to_string(), ts),
+            ],
+        )
+        .unwrap();
+
+        assert_eq!(
+            info.to_file_content(),
+            format!(
+                "follower=127.0.0.1:8000,{TEST_TIME_STAMP},node1\n\
+                 follower=127.0.0.1:8001,{TEST_TIME_STAMP},node2"
+            ),
+        );
+    }
+
+    #[test]
+    fn cluster_empty() {
+        let info = NodeDiscovery::cluster(None, vec![]).unwrap();
+        assert_eq!(info.to_file_content(), "");
+    }
+
+    #[test]
+    fn cluster_errors_on_duplicate_node_id() {
+        let ts = test_timestamp();
+        let result = NodeDiscovery::cluster(
+            None,
+            vec![
+                ("node1".to_string(), "127.0.0.1:8000".to_string(), ts),
+                ("node1".to_string(), "127.0.0.1:8001".to_string(), ts),
+            ],
+        );
+
+        let err = result.unwrap_err();
+        assert!(err
+            .to_string()
+            .contains("Duplicate node id found in Nodes table: node1"));
+    }
+
+    #[test]
+    fn cluster_errors_when_leader_not_in_nodes() {
+        let ts = test_timestamp();
+        let result = NodeDiscovery::cluster(
+            Some("missing_leader".to_string()),
+            vec![
+                ("node1".to_string(), "127.0.0.1:8000".to_string(), ts),
+                ("node2".to_string(), "127.0.0.1:8001".to_string(), ts),
+            ],
+        );
+
+        let err = result.unwrap_err();
+        assert!(err
+            .to_string()
+            .contains("Leader is missing from the Nodes table."));
+    }
+
+    #[test]
+    fn leader_change_detected_when_membership_unchanged() {
+        let ts = test_timestamp();
+        let nodes = vec![
+            ("node1".to_string(), "127.0.0.1:8000".to_string(), ts),
+            ("node2".to_string(), "127.0.0.1:8001".to_string(), ts),
+        ];
+
+        // Initial state: node1 is leader
+        let info1 = NodeDiscovery::cluster(Some("node1".to_string()), nodes.clone()).unwrap();
+
+        // New state: node2 becomes leader, same membership
+        let info2 = NodeDiscovery::cluster(Some("node2".to_string()), nodes).unwrap();
+
+        // Followers differ because the previous leader becomes a follower and vice versa.
+        assert_ne!(info1.followers, info2.followers);
+
+        // But leader_id should differ.
+        assert_eq!(info1.leader_id(), Some("node1".to_string()));
+        assert_eq!(info2.leader_id(), Some("node2".to_string()));
+    }
+}
