+++ examples/demo-rollup/tests/replica/replica_registers_in_db.rs
@@ -21,17 +21,13 @@ async fn test_multiple_replicas_register_in_nodes_table() {
         .await;
 
     // Verify both replicas are present in followers
-    let cluster_info = setup.cluster_info_subscription.wait_for_change().await;
+    let cluster_info = setup.wait_for_cluster_change().await;
 
     assert_eq!(cluster_info.followers.len(), 2);
     assert!(cluster_info.has_follower("replica_then_leader"));
     assert!(cluster_info.has_follower("replica"));
 
-    let old_follower = cluster_info
-        .followers
-        .iter()
-        .find(|f| f.node_id == "replica")
-        .unwrap();
+    let old_follower = cluster_info.followers.get("replica").unwrap();
 
     let mut builder = replica_then_leader.shutdown().await.unwrap();
     // Upgrade to leader.
@@ -40,17 +36,13 @@ async fn test_multiple_replicas_register_in_nodes_table() {
     let leader = builder.start_test_rollup().await.unwrap();
     leader.wait_for_sequencer_ready().await.unwrap();
 
-    let new_cluster_info = setup.cluster_info_subscription.wait_for_change().await;
+    let new_cluster_info = setup.wait_for_cluster_change().await;
 
     assert!(new_cluster_info.has_leader("replica_then_leader"));
     assert!(new_cluster_info.has_follower("replica"));
 
     // Verify that nodes timestamps are increasing.
-    let new_follower = new_cluster_info
-        .followers
-        .iter()
-        .find(|f| f.node_id == "replica")
-        .unwrap();
+    let new_follower = new_cluster_info.followers.get("replica").unwrap();
 
     assert!(new_follower.last_updated > old_follower.last_updated);
 
@@ -84,7 +76,7 @@ async fn test_stale_nodes_are_filtered_from_cluster_info() {
         .await;
 
     // Wait for both nodes to appear in cluster info.
-    let cluster_info = setup.cluster_info_subscription.wait_for_change().await;
+    let cluster_info = setup.wait_for_cluster_change().await;
     assert!(
         cluster_info.has_leader("leader"),
         "Leader should be present"
@@ -99,8 +91,7 @@ async fn test_stale_nodes_are_filtered_from_cluster_info() {
 
     // Wait for the replica to become stale and be filtered out.
     let cluster_info = setup
-        .cluster_info_subscription
-        .wait_for_change_with_timeout(Duration::from_secs(10))
+        .wait_for_cluster_change_with_timeout(Duration::from_secs(10))
         .await;
 
     // Leader should always be present (never filtered regardless of age).
@@ -109,9 +100,10 @@ async fn test_stale_nodes_are_filtered_from_cluster_info() {
         "Leader should always be present in cluster info"
     );
 
-    // When there are no other followers, the leader is added to the followers list
-    // so that clients always have at least one node to connect to for reads.
-    assert!(cluster_info.has_follower("leader"));
+    // With follower storage keyed by follower node IDs, no follower should remain
+    // once the replica becomes stale.
+    assert!(cluster_info.followers.is_empty());
+    assert!(!cluster_info.has_follower("leader"));
     assert!(!cluster_info.has_follower("replica"));
 
     let _ = leader.shutdown().await;
